requirements.txt
# MathemaTest Phase 1: Ingestion Pipeline Dependencies
# Requires Python 3.10+ and CUDA 11.8+

# Layout Detection
doclayout-yolo>=0.0.2

# Formula OCR
unimernet>=0.2.2
transformers>=4.35.0
accelerate>=0.25.0
sentencepiece>=0.1.99
huggingface-hub>=0.19.0

# Symbolic Math
sympy>=1.12
latex2sympy2>=1.9.1

# Data Modeling
pydantic>=2.5.0

# PDF Processing
pdf2image>=1.16.0
opencv-python>=4.8.0
pillow>=9.5.0

# Core Dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# ==========================================
# Phase 2-3: Knowledge Base & Hybrid Retrieval
# ==========================================

# Knowledge Graph
neo4j>=5.14.0
python-dotenv>=1.0.0

# Vector Store
chromadb>=0.4.18
sentence-transformers>=2.2.2

# LLM & Retrieval
openai>=1.6.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-openai>=0.0.5

# Re-ranking
# cross-encoder model loaded via sentence-transformers

# System dependency (required for pdf2image):
# sudo apt-get install poppler-utils  # Ubuntu/Debian
# brew install poppler                 # macOS
